scrapers/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
anthropic==0.39.0
pdfplumber==0.11.4
Pillow==10.4.0
requests==2.31.0
python-dotenv==1.0.0
supabase==2.9.1

